+++ demo/router/routes.js
@@ -26,6 +26,7 @@ export default [
   page('Modals'),
   page('Icons'),
   page('Other'),
+  page('Logger'),
   page('Testing'),
   {
     path: '*',