+++ src/constants/index.js
@@ -1,4 +1,17 @@
+import { createScssVariablesDictionary } from '@/utils/styles';
+import variables from '@/styles/export.scss';
+
 /**
- * @type {String}
+ * @type {Object}
  */
-export const QA_ATTRIBUTE_NAME = 'data-qa';
+export const STYLE_VARIABLES = Object.freeze(createScssVariablesDictionary(variables));
+
+/**
+ * @type {Readonly<Object>}
+ */
+export const BREAKPOINTS = Object.freeze(STYLE_VARIABLES.breakpoints);
+
+/**
+ * @type {Readonly<Object>}
+ */
+export const COLORS = Object.freeze(STYLE_VARIABLES.colors);