+++ src/utils/errors/InvalidTypeError.js
@@ -13,7 +13,7 @@ export default class InvalidTypeError extends BaseError {
    * throw new InvalidTypeError(str, 'str', 'String');
    */
   constructor(value, name, type) {
-    const message = `Invalid type for argument "${name}", expected "${type}" got "${typeof value}"`;
+    const message = `Invalid type for argument "${name}", expected "${type}", got "${typeof value}"`;
 
     super(message);
   }