+++ src/utils/helpers/getSlot.js
@@ -1,16 +1,15 @@
 /**
- * @param {Object<VueInstance>} parent
+ * @param {Object<Vue>} context
  * @param {String} slotName
  * @param {Object} scope
  * @returns {*}
  */
-export default function getSlot(parent, slotName = 'default', scope = {}) {
-  const slot = parent.$slots[slotName];
-  const scopedSlot = parent.$scopedSlots[slotName];
+export default function getSlot(context, slotName = 'default', scope = {}) {
+  const slot = context.$slots[slotName];
 
-  if (slot) {
-    return slot;
-  }
+  if (slot) return slot;
+
+  const scopedSlot = context.$scopedSlots[slotName];
 
   return scopedSlot ? scopedSlot(scope) : null;
 }