+++ src/utils/index.js
@@ -1,6 +1,10 @@
+export { default as archiveFilesToZip } from './archiveFilesToZip';
 export { default as asyncDelay } from './asyncDelay';
 export { default as cloneArray } from './cloneArray';
 export { default as cloneDeep } from './cloneDeep';
+export { default as downloadFile } from './downloadFile';
+export { default as downloadFileToBrowser } from './downloadFileToBrowser';
+export { default as freezeDeep } from './freezeDeep';
 export { default as generateHash } from './generateHash';
 export { default as isEqual } from './isEqual';
 export { default as isEven } from './isEven';
@@ -14,6 +18,7 @@ export { default as pascalToKebab } from './pascalToKebab';
 export { default as percentOfInteger } from './percentOfInteger';
 export { default as randomArrayElement } from './randomArrayElement';
 export { default as randomInteger } from './randomInteger';
+export { default as removeDomElement } from './removeDomElement';
 export { default as scrollTo } from './scrollTo';
 export { default as shuffle } from './shuffle';
 export { default as throttle } from './throttle';