+++ src/utils/inspect/index.js
@@ -1,7 +1,9 @@
 export { default as isArray } from './isArray';
 export { default as isBoolean } from './isBoolean';
+export { default as isElement } from './isElement';
 export { default as isEmptyObject } from './isEmptyObject';
 export { default as isFunction } from './isFunction';
+export { default as isInteger } from './isInteger';
 export { default as isNaN } from './isNaN';
 export { default as isNil } from './isNil';
 export { default as isNull } from './isNull';
@@ -9,4 +11,5 @@ export { default as isNumber } from './isNumber';
 export { default as isObjectLike } from './isObjectLike';
 export { default as isPlainObject } from './isPlainObject';
 export { default as isString } from './isString';
+export { default as isSymbol } from './isSymbol';
 export { default as isUndefined } from './isUndefined';