+++ src/utils/styles/getScssVariable.js
@@ -1,9 +1,18 @@
-import variables from '@/styles/export.scss';
+import { InvalidTypeError } from '../errors';
+import { isString, isPlainObject } from '../inspect';
 
 /**
+ * @param {Object} dictionary Словарь переменных
  * @param {String} variableKey Имя переменной
  * @returns {Object}
  */
-export default function getScssVariable(variableKey) {
-  return variables[variableKey];
+export default function getScssVariable(dictionary, variableKey) {
+  if (!isPlainObject(dictionary)) {
+    throw new InvalidTypeError(dictionary, 'dictionary', 'Object');
+  }
+  if (!isString(variableKey)) {
+    throw new InvalidTypeError(variableKey, 'variableKey', 'String');
+  }
+
+  return dictionary[variableKey];
 }