+++ src/utils/styles/getScssVariableByMap.js
@@ -1,10 +1,24 @@
-import variables from '@/styles/export.scss';
+import { InvalidTypeError } from '../errors';
+import { isString, isPlainObject } from '../inspect';
+
+const SEPARATOR = '-';
 
 /**
+ * @param {Object} dictionary Словарь переменных
  * @param {String} mapKey Префикс переменной, на основании которого будет сформирована карта
  * @param {String} variableKey Имя переменной
  * @returns {*}
  */
-export default function getScssVariableByMap(mapKey, variableKey) {
-  return variables[`${mapKey}-${variableKey}`];
+export default function getScssVariableByMap(dictionary, mapKey, variableKey) {
+  if (!isPlainObject(dictionary)) {
+    throw new InvalidTypeError(dictionary, 'dictionary', 'Object');
+  }
+  if (!isString(mapKey)) {
+    throw new InvalidTypeError(mapKey, 'mapKey', 'String');
+  }
+  if (!isString(variableKey)) {
+    throw new InvalidTypeError(variableKey, 'variableKey', 'String');
+  }
+
+  return dictionary[`${mapKey}${SEPARATOR}${variableKey}`];
 }